requirements.txt
biopython==1.76
Click==7.0
cycler==0.10.0
DendroPy==4.4.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
